Fix CrunchBang detection, gsettings error output.
I'm not sure why the release file had its name changed in Waldorf, but screenFetch currently identifies #! as Debian as a result.

Single-bracket conditionals always evaluate to true without an argument, which is the case when `which` doesn't find anything. At present, this causes screenFetch to output a nasty `gsettings: command not found` message on #!. Using double brackets instead prevents this.