Missed a set of parentheses in the pushed version
I tested the code under cygwin with one monitor, and found that even that produced wrong output (and a test with some faked monitors confirmed the output given in the issue). The error was me omitting the parentheses that signify an array in bash. With them there it should work as intended.