Changed recommended installation command to be apt
I added a note about Debian or Ubuntu 14.04 vs Ubuntu 16.04 when it comes to what command is used to install things. As of Ubuntu 16.04, `apt` is the recommended command, rather than `apt-get`. I also added `sudo` because you need to include that if you're not running a root terminal (and it doesn't hurt anything even if you are running a root terminal).

I also added newlines to the beginning and end just because I like newlines.