Update the Parabola logo
One of our users posted a screenshot made with screenFetch, and I noticed
that the logo was based off of the old pre-2013 version of our logo (it had
a shadow, the new one has some slashes through it).

Anyway, I've got an open PR against archey3 to add our logo; so I simply
transferred over the ASCII Parabola logo that I made for that.  I made it
by opening our logo in Inkscape, putting a textbox over it, and fiddling
with characters until I thought it did a good job of covering the logo.