Fix detection of Kali Linux when lsb_release is not installed
When lsb_release is not available (eg. when running Kali Linux in a
container), detection is done based on the ID variable from the
os-release file.  For Kali Linux, this ID is simply 'kali'.

However screenfetch is looking for the glob 'kali*linux', so it doesn't
work:

  ┌──(like㉿kali)-[/work]
  └─# screenfetch -vvvvv
  :: Finding distro...found as 'kali '

  [...]

                               OS: kali
                               Kernel: xxx
           #####               Uptime: xxx
          #######              Packages: Unknown
          ##O#O##              Shell: xxx
          #######              Disk: xxx
        ###########            CPU: xxx
       #############           GPU: xxx
      ###############          RAM: xxx
      ################
     #################
   #####################
   #####################
     #################

Fixed with this commit.