Fix RAM information not displaying correctly
When running screenmatch when the system language is Chinese, an error will be reported and RAM information cannot be displayed correctly. The reason is that when the system language is Chinese, the colon used to execute "free -b" output is a Chinese colon, and the characters cannot be correctly segmented.